Refactor Alpaca config into ES module
