+++ server/src/trade/broker/providers/alpaca/alpaca.config.js
@@ -1,0 +1,4 @@
+export default {
+    liveApiBaseUrl: 'https://api.alpaca.markets',
+    testApiBaseUrl: 'https://paper-api.alpaca.markets'
+};
