--- server/src/trade/broker/providers/alpaca/alpaca.config.json
@@ -1,4 +1,0 @@
-{
-    "liveApiBaseUrl": "https://api.alpaca.markets",
-    "testApiBaseUrl": "https://paper-api.alpaca.markets"
-}
